pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Speaker count estimation with single speakers"
authors = ["Aucke <aucke.bos97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.3"
scipy = "^1.6.1"
pydub = "^0.25.1"
scikit-learn = "^0.24.1"
tensorflow-gpu = "^2.4.1"
stft = "^0.5.2"
librosa = "^0.8.0"
python_speech_features = "^0.6"
tensorflow-probability = "^0.12.1"
matplotlib = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
